Remove and ignore package-lock.json
Lockfiles are only useful for people developing on this project,
they're not relevant to the published package (where only
the version ranges described in package.json are used to choose
dependencies).

Our standard practice elsewhere is to only include `yarn.lock`,
and ignore `package-lock.json` so it never gets added.

This is because, while we have both, one or the other will probably
get wildly out of date, and so we'll be describing two different sets
of dependencies, only one of which (Yarn) has actually been tested.